feat: Correção

+++ ex6.js
@@ -1,7 +1,7 @@
 // Escreva um programa que solicita o ano de nascimento de uma pessoa e verifica se ela é maior
 // de 18 anos. Exiba uma mensagem indicando se a pessoa é maior de idade ou não.
 
-let data = parseFloat (prompt("Digite sua data de nascimento:"))
+let data = parseInt(prompt("Digite sua data de nascimento:"))
 
 let idade = (2024 - data)
